# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PaperMatrix API
  version: 0.1.0
  description: Local-first file-backed research paper management API.
servers:
  - url: http://127.0.0.1:8000/api/v1
paths:
...
  /projects/{project_id}/papers/{paper_id}/note:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaperNote
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperNote" }
    put:
      operationId: updatePaperNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperNoteUpdate"
      responses:
        "200":
          description: Saved note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperNote" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/supplement:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaperSupplement
      summary: 读取独立的个人补充笔记
      responses:
        "200":
          description: Supplement note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperNote" }
    put:
      operationId: updatePaperSupplement
      summary: 使用独立 revision 保存个人补充笔记
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaperNoteUpdate" }
      responses:
        "200":
          description: Saved supplement note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperNote" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/items:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaperNoteItems
      summary: 读取所有模板槽位（含空项）、可扩展条目和证据目录
      responses:
        "200":
          description: Note item document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteItemDocument" }
    post:
      operationId: createPaperNoteItem
      summary: 在允许扩展的模板组内新增独立条目
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteItemCreate" }
      responses:
        "201":
          description: Created note item and synchronized projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteItemUpdateResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/slots/{template_key}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
      - in: path
        name: template_key
        required: true
        schema: { type: string, maxLength: 40 }
    put:
      operationId: updatePaperNoteSlot
      summary: 在模板原位置填写或清空一个固定槽位
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteSlotUpdate" }
      responses:
        "200":
          description: Updated fixed slot, note and analysis projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteSlotUpdateResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/evidence:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: createPaperEvidence
      summary: 新增论文级证据并可同时关联当前模板条目
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvidenceCreate" }
      responses:
        "201":
          description: Created evidence, updated note and analysis projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceCreateResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/items/{item_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
      - in: path
        name: item_id
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: updatePaperNoteItem
      summary: 只替换稳定锚点对应的 Markdown 片段并同步条目投影
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteItemUpdate" }
      responses:
        "200":
          description: Updated note and analysis projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteItemUpdateResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/items/{item_id}/favorite:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
      - in: path
        name: item_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updatePaperNoteItemFavorite
      summary: 更新确认条目的重点收藏状态
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteItemFavoriteUpdate" }
      responses:
        "200":
          description: Updated favorite state and analysis projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteItemFavoriteUpdateResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/note/items/delete:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: deletePaperNoteItems
      summary: 批量删除条目对应的 Markdown 标题块和分析投影
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteItemDeleteRequest" }
      responses:
        "200":
          description: Updated note and analysis projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteItemDeleteResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/questions:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    get:
      operationId: getPaperQuestions
      summary: 获取论文问题清单；未保存时 revision 为 0
      responses:
        "200":
          description: Questions document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionsDocument" }
    post:
      operationId: createPaperQuestion
      summary: 添加问题，答案可以为空
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuestionInput" }
      responses:
        "201":
          description: Updated questions document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionsDocument" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/questions/{question_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
      - in: path
        name: question_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updatePaperQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuestionInput" }
      responses:
        "200":
          description: Updated questions document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionsDocument" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deletePaperQuestion
      parameters:
        - in: query
          name: expected_revision
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Updated questions document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionsDocument" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
...
  /projects/{project_id}/papers/{paper_id}/analysis/parse-note:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: previewPaperNoteAnalysis
      summary: 只读解析已保存的结构化笔记并生成待审阅候选
      responses:
        "200":
          description: Candidate preview without authoritative file changes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteParsePreview" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
  /projects/{project_id}/papers/{paper_id}/analysis/import-candidates:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/PaperId"
    post:
      operationId: importPaperNoteAnalysisCandidates
      summary: 确认选中候选并一次性写入论文分析条目
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateImportRequest" }
      responses:
        "200":
          description: Import result and updated analysis document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateImportResult" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
        "422": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  parameters:
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
    PaperId:
      in: path
      name: paper_id
      required: true
      schema: { type: string, format: uuid }
  responses:
    ErrorResponse:
      description: PaperMatrix error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            action: { type: [string, "null"] }
            request_id: { type: string }
...
    PaperNote:
      type: object
      required: [paper_id, markdown, revision, updated_at]
      properties:
        paper_id: { type: string, format: uuid }
        markdown: { type: string }
        revision: { type: integer, minimum: 0 }
        updated_at: { type: string, format: date-time }
    PaperNoteUpdate:
      type: object
      required: [markdown, expected_revision]
      properties:
        markdown: { type: string }
        expected_revision: { type: integer, minimum: 0 }
    EvidenceReference:
      type: object
      required:
        [
          paper_id,
          evidence_code,
          page_label,
          pdf_page_index,
          section,
          figure,
          table,
          locator_note,
        ]
      properties:
        evidence_id: { type: string, format: uuid }
        evidence_code: { type: [string, "null"], maxLength: 40 }
        paper_id: { type: string, format: uuid }
        page_label: { type: [string, "null"] }
        pdf_page_index: { type: [integer, "null"], minimum: 1 }
        section: { type: [string, "null"] }
        figure: { type: [string, "null"] }
        table: { type: [string, "null"] }
        locator_note: { type: string }
...
    QuestionInput:
      type: object
      required: [question, status, answer, evidence, tags, expected_revision]
      properties:
        question: { type: string, minLength: 1, maxLength: 5000 }
        status: { type: string, enum: [open, answered, deferred] }
        answer: { type: string, maxLength: 50000 }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/EvidenceReference" }
        tags:
          type: array
          items: { type: string }
        expected_revision: { type: integer, minimum: 0 }
    QuestionsDocument:
      type: object
      required: [schema_version, paper_id, revision, updated_at, questions]
      properties:
        schema_version: { type: integer, const: 2 }
        paper_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        updated_at: { type: string, format: date-time }
        questions:
          type: array
          items: { $ref: "#/components/schemas/ReadingQuestion" }
...
    NoteParsePreview:
      type: object
      required: [paper_id, note_revision, paper_revision, candidates, removals, warnings]
      properties:
        paper_id: { type: string, format: uuid }
        note_revision: { type: integer, minimum: 0 }
        paper_revision: { type: integer, minimum: 1 }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/NoteAnalysisCandidate" }
        removals:
          type: array
          items: { $ref: "#/components/schemas/NoteAnalysisRemoval" }
        warnings:
          type: array
          items: { type: string }
    CandidateImportRequest:
      type: object
      required: [expected_note_revision, expected_paper_revision]
      anyOf:
        - required: [candidate_ids]
          properties:
            candidate_ids: { minItems: 1 }
        - required: [removal_item_ids]
          properties:
            removal_item_ids: { minItems: 1 }
      properties:
        candidate_ids:
          type: array
          default: []
          items: { type: string, format: uuid }
        removal_item_ids:
          type: array
          default: []
          items: { type: string, format: uuid }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
    CandidateImportResult:
      type: object
      required:
        [
          analysis,
          note,
          imported_items,
          synchronized_items,
          skipped_candidate_ids,
          superseded_item_ids,
          deleted_item_ids,
        ]
      properties:
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        note: { $ref: "#/components/schemas/PaperNote" }
        imported_items:
          type: array
          items: { $ref: "#/components/schemas/AnalysisItem" }
        synchronized_items:
          type: array
          items: { $ref: "#/components/schemas/AnalysisItem" }
        skipped_candidate_ids:
          type: array
          items: { type: string, format: uuid }
        superseded_item_ids:
          type: array
          items: { type: string, format: uuid }
        deleted_item_ids:
          type: array
          items: { type: string, format: uuid }
...
    NoteItemSource:
      type: object
      required:
        [
          item_id,
          kind,
          display_label,
          title,
          section_key,
          section_title,
          section_order,
          markdown,
          source_fingerprint,
          sync_status,
          is_favorite,
        ]
      properties:
        item_id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/AnalysisItem/properties/kind" }
        display_label: { type: [string, "null"], maxLength: 80 }
        title: { type: string }
        section_key: { type: [string, "null"] }
        section_title: { type: [string, "null"] }
        section_order: { type: [integer, "null"], minimum: 1 }
        markdown: { type: string }
        source_fingerprint:
          type: [string, "null"]
          pattern: "^[0-9a-f]{64}$"
        sync_status:
          type: string
          enum: [synced, review_required, missing]
        is_favorite: { type: boolean }
    NoteItemDocument:
      type: object
      required:
        [
          paper_id,
          note_revision,
          paper_revision,
          item_templates,
          slots,
          evidence_catalog,
          items,
          candidates,
          removals,
          warnings,
          pending_candidate_count,
        ]
      properties:
        paper_id: { type: string, format: uuid }
        note_revision: { type: integer, minimum: 0 }
        paper_revision: { type: integer, minimum: 1 }
        item_templates:
          type: array
          items: { $ref: "#/components/schemas/NoteItemTemplate" }
        slots:
          type: array
          items: { $ref: "#/components/schemas/NoteItemSlot" }
        evidence_catalog:
          type: array
          items: { $ref: "#/components/schemas/EvidenceReference" }
        items:
          type: array
          items: { $ref: "#/components/schemas/NoteItemSource" }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/NoteAnalysisCandidate" }
        removals:
          type: array
          items: { $ref: "#/components/schemas/NoteAnalysisRemoval" }
        warnings:
          type: array
          items: { type: string }
        pending_candidate_count: { type: integer, minimum: 0 }
    NoteItemTemplate:
      type: object
      required:
        [
          template_key,
          chapter,
          kind,
          label,
          description,
          heading,
          heading_level,
          repeatable,
          child_heading_prefix,
          insert_before_heading,
          body_template,
        ]
      properties:
        template_key: { type: string, maxLength: 40 }
        chapter: { type: integer, minimum: 1, maximum: 8 }
        kind: { $ref: "#/components/schemas/AnalysisItem/properties/kind" }
        label: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 300 }
        heading: { type: string, maxLength: 300 }
        heading_level: { type: integer, enum: [2, 3, 4] }
        repeatable: { type: boolean }
        child_heading_prefix: { type: string, maxLength: 80 }
        insert_before_heading: { type: [string, "null"], maxLength: 300 }
        body_template: { type: string, maxLength: 20000 }
    NoteItemSlot:
      type: object
      required:
        [
          slot_key,
          template_key,
          kind,
          label,
          description,
          section_title,
          markdown,
          item_id,
          source_fingerprint,
          sync_status,
          is_favorite,
          repeatable,
          repeatable_template_key,
          can_delete,
        ]
      properties:
        slot_key: { type: string, maxLength: 80 }
        template_key: { type: string, maxLength: 40 }
        kind: { $ref: "#/components/schemas/AnalysisItem/properties/kind" }
        label: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 300 }
        section_title: { type: string, maxLength: 300 }
        markdown: { type: string, maxLength: 200000 }
        item_id: { type: [string, "null"], format: uuid }
        source_fingerprint:
          type: [string, "null"]
          pattern: "^[0-9a-f]{64}$"
        sync_status:
          type: string
          enum: [empty, synced, review_required, missing]
        is_favorite: { type: boolean }
        repeatable: { type: boolean }
        repeatable_template_key: { type: [string, "null"], maxLength: 40 }
        can_delete: { type: boolean }
    NoteItemCreate:
      type: object
      required:
        [
          template_key,
          title,
          expected_note_revision,
          expected_paper_revision,
        ]
      properties:
        template_key: { type: string, minLength: 1, maxLength: 40 }
        title: { type: string, minLength: 1, maxLength: 300 }
        markdown: { type: string, maxLength: 200000 }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
    NoteSlotUpdate:
      type: object
      required: [markdown, expected_note_revision, expected_paper_revision]
      properties:
        markdown: { type: string, maxLength: 200000 }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
    EvidenceCreate:
      type: object
      required:
        [
          locator_note,
          expected_note_revision,
          expected_paper_revision,
        ]
      properties:
        item_id: { type: [string, "null"], format: uuid }
        evidence_type: { type: string, maxLength: 80 }
        page_label: { type: [string, "null"], maxLength: 40 }
        pdf_page_index: { type: [integer, "null"], minimum: 1 }
        section: { type: [string, "null"], maxLength: 200 }
        figure: { type: [string, "null"], maxLength: 80 }
        table: { type: [string, "null"], maxLength: 80 }
        locator_note: { type: string, minLength: 1, maxLength: 2000 }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
    EvidenceCreateResult:
      type: object
      required: [note, analysis, evidence, item]
      properties:
        note: { $ref: "#/components/schemas/PaperNote" }
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        evidence: { $ref: "#/components/schemas/EvidenceReference" }
        item:
          oneOf:
            - $ref: "#/components/schemas/AnalysisItem"
            - type: "null"
    NoteItemUpdate:
      type: object
      required:
        [
          markdown,
          expected_note_revision,
          expected_paper_revision,
          expected_source_fingerprint,
        ]
      properties:
        markdown: { type: string, minLength: 1, maxLength: 200000 }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
        expected_source_fingerprint:
          type: string
          pattern: "^[0-9a-f]{64}$"
    NoteItemUpdateResult:
      type: object
      required: [note, analysis, item]
      properties:
        note: { $ref: "#/components/schemas/PaperNote" }
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        item: { $ref: "#/components/schemas/AnalysisItem" }
    NoteSlotUpdateResult:
      type: object
      required: [note, analysis, slot, item]
      properties:
        note: { $ref: "#/components/schemas/PaperNote" }
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        slot: { $ref: "#/components/schemas/NoteItemSlot" }
        item:
          oneOf:
            - $ref: "#/components/schemas/AnalysisItem"
            - type: "null"
    NoteItemFavoriteUpdate:
      type: object
      required: [is_favorite, expected_paper_revision]
      properties:
        is_favorite: { type: boolean }
        expected_paper_revision: { type: integer, minimum: 1 }
    NoteItemFavoriteUpdateResult:
      type: object
      required: [analysis, item]
      properties:
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        item: { $ref: "#/components/schemas/AnalysisItem" }
    NoteItemDeleteRequest:
      type: object
      required: [item_ids, expected_note_revision, expected_paper_revision]
      properties:
        item_ids:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
        expected_note_revision: { type: integer, minimum: 0 }
        expected_paper_revision: { type: integer, minimum: 1 }
    NoteItemDeleteResult:
      type: object
      required: [note, analysis, deleted_item_ids]
      properties:
        note: { $ref: "#/components/schemas/PaperNote" }
        analysis: { $ref: "#/components/schemas/PaperAnalysisDocument" }
        deleted_item_ids:
          type: array
          items: { type: string, format: uuid }
...